Create P4P3 (Two cylinders)
Write a program that asks the user to input the values of 2 cylinders’ radius (float) & height (float). Write a function (cylinder) that accepts two parameters (r & h). The function shall return the volume of both cylinders rounded to 3 decimal places:

Example:
------------------------------
Please enter the radius of a cylinder 1: 5
Please enter the height of a cylinder 1: 3.2
Please enter the radius of a cylinder 2: 2.1
Please enter the height of a cylinder 2: 8.2
#Using the same function should result into:
The volume of the cylinder 1 is: 251.327 cubic units.
The volume of the cylinder 2 is: 113.606 cubic units.