Create P4P4
Write a program that asks the user to enter the number of circles to draw. The program should ask the user for the coordinates (x,y), size, and color for every circle. The program will use Python Turtle to print out the circles. Hint: there should be a function (draw_circle) within your program that accepts 4 parameters (x, y, size, color) to draw a circle.