Create P2P2 (USD to MXN)
Write a program that converts the currency between U.S. Dollars (USD) and Mexican Pesos (MXN). First, prompt the user to enter the exchange rate. Second, prompt the user to enter 0 to convert from dollars to pesos and 1 for vice versa. Finally, enter the amount to be converted. The program should print the value of the converted currency.

Example output:
-------------------
Enter the exchange rate from USD to MXN: 20.32
Enter 0 to convert USD to MXN, or 1 vice versa: 1
Enter the MXN amount: 1000
1000 MXN equals 49.22 USD